refactor: move author enrichment helpers to _book_enrichment.py
Moves _resolve_author_keys, _extract_author_keys, and
enrich_authors_from_work to _book_enrichment.py so that _enrich_one
also enriches authors during the paper enrichment pipeline. Fixes
ol_client fixture usage in get_author tests.
