+++ src/scholar_mcp/_book_enrichment.py
@@ -15,6 +15,71 @@ from ._server_deps import ServiceBundle
 logger = logging.getLogger(__name__)
 
 
+async def _resolve_author_keys(
+    author_keys: list[str], bundle: ServiceBundle
+) -> list[str]:
+    """Resolve Open Library author keys to author names.
+
+    Args:
+        author_keys: List of Open Library author keys (e.g. ``/authors/OL239963A``).
+        bundle: Service bundle with openlibrary client.
+
+    Returns:
+        List of resolved author name strings.
+    """
+    # Strip path prefix — get_author expects short IDs like "OL239963A"
+    ids = [k.rsplit("/", 1)[-1] for k in author_keys]
+    results = await asyncio.gather(
+        *(bundle.openlibrary.get_author(aid) for aid in ids)
+    )
+    return [a["name"] for a in results if a and a.get("name")]
+
+
+def _extract_author_keys(work: dict[str, Any]) -> list[str]:
+    """Extract author keys from a work record.
+
+    Args:
+        work: Open Library work dict.
+
+    Returns:
+        List of author key strings.
+    """
+    keys: list[str] = []
+    for entry in work.get("authors") or []:
+        if isinstance(entry, dict):
+            author = entry.get("author")
+            if isinstance(author, dict) and author.get("key"):
+                keys.append(author["key"])
+    return keys
+
+
+async def enrich_authors_from_work(book: BookRecord, bundle: ServiceBundle) -> None:
+    """Enrich book in-place with authors from its work record.
+
+    Best-effort: failures are logged and silently skipped.
+
+    Args:
+        book: Book record to enrich in-place.
+        bundle: Service bundle with openlibrary client.
+    """
+    if book.get("authors"):
+        return
+    work_id = book.get("openlibrary_work_id")
+    if not work_id:
+        return
+    try:
+        work = await bundle.openlibrary.get_work(work_id)
+        if work is None:
+            return
+        author_keys = _extract_author_keys(work)
+        if author_keys:
+            names = await _resolve_author_keys(author_keys, bundle)
+            if names:
+                book["authors"] = names
+    except Exception:
+        logger.debug("author_enrichment_failed work_id=%s", work_id, exc_info=True)
+
+
 def _needs_book_enrichment(paper: dict[str, Any]) -> bool:
     """Check whether a paper should be enriched with book metadata.
 
@@ -66,6 +131,7 @@ async def _enrich_one(paper: dict[str, Any], bundle: ServiceBundle) -> None:
             return
 
         book: BookRecord = normalize_book(edition, source="edition")
+        await enrich_authors_from_work(book, bundle)
         await bundle.cache.set_book_by_isbn(isbn, book)
         work_id = book.get("openlibrary_work_id")
         if work_id:
